Origin/feature/it21365614 refine

+++ src/main/java/com/example/onlineticketingsystem/controller/BusController.java
@@ -3,6 +3,9 @@ package com.example.onlineticketingsystem.controller;
 import com.example.onlineticketingsystem.DTO.BusDTO;
 import com.example.onlineticketingsystem.service.BusService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -11,27 +14,49 @@ import java.util.List;
 @RequestMapping(value = "api/v1/bus")
 @CrossOrigin
 public class BusController {
+
     @Autowired
     private BusService busService;
 
+//    @GetMapping
+//    public ResponseEntity<?> helloBus(){
+//        return ResponseEntity.ok("Hello Bus");
+//    }
+
     @GetMapping("/getBus")
-    public List<BusDTO> getBus(){
-        return busService.getAllBuses();
+    public ResponseEntity<?> getBus(Authentication authentication){
+        if (authentication.getAuthorities().stream().noneMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("admin") )){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. You need to have admin role.");
+        }
+        List<BusDTO> busDTOList = busService.getAllBuses();
+        return ResponseEntity.ok(busDTOList);
     }
 
     @PutMapping("/updateBus")
-    public BusDTO updateBus(@RequestBody BusDTO busDTO){
-        return busService.updateBus(busDTO);
+    public ResponseEntity<?> updateBus(@RequestBody BusDTO busDTO, Authentication authentication){
+        if (authentication.getAuthorities().stream().noneMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("admin") || grantedAuthority.getAuthority().equals("bus-owner"))){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. You need to have admin Or bus-owner role.");
+        }
+        BusDTO bus = busService.updateBus(busDTO);
+        return ResponseEntity.ok(bus);
     }
 
     @DeleteMapping("/deleteBus")
-    public boolean deleteBus(@RequestBody BusDTO busDTO){
-        return busService.deleteBus(busDTO);
+    public ResponseEntity<?> deleteBus(@RequestBody BusDTO busDTO, Authentication authentication){
+        if (authentication.getAuthorities().stream().noneMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("admin") || grantedAuthority.getAuthority().equals("bus-owner"))){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. You need to have admin Or bus-owner role.");
+        }
+        Boolean isDeleted = busService.deleteBus(busDTO);
+        return ResponseEntity.ok(isDeleted);
     }
 
     @PostMapping("/saveBus")
-    public BusDTO saveBus(@RequestBody BusDTO busDTO){
-        return busService.saveBus(busDTO);
+    public ResponseEntity<?> saveBus(@RequestBody BusDTO busDTO, Authentication authentication){
+        if (authentication.getAuthorities().stream().noneMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("admin") || grantedAuthority.getAuthority().equals("bus-owner"))){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Access denied. You need to have admin Or bus-owner role.");
+        }
+        BusDTO bus = busService.saveBus(busDTO);
+        return ResponseEntity.ok(bus);
     }
